cryptohome: Support for get_enrollment_id action with ignore_cache option
The get_enrollment_id action from cryptohome caches the result first time
the EID is computed and uses the value from cache to return the data. This
change will allow the caller to specify to ignore the cache. This means the
EID should be recomputed while the new value should not be stored in cache.

BUG=chromium:835759
TEST=unit test and manual test with cryptohome --action=get_enrollment_id --ignore_cache
CQ-DEPEND=CL:1024031

Change-Id: I2659f4b775b2de98d7592e793c186063ce78823c
Reviewed-on: https://chromium-review.googlesource.com/1024030
Commit-Ready: Igor <igorcov@chromium.org>
Tested-by: Igor <igorcov@chromium.org>
Reviewed-by: Igor <igorcov@chromium.org>
Reviewed-by: Yves Arrouye <drcrash@chromium.org>
Reviewed-by: Andrey Pronin <apronin@chromium.org>