attestation: Support for get_enrollment_id action with ignore_cache option
The get_enrollment_id action from cryptohome caches the result first time
the EID is computed and uses the value from cache to return the data. This
change will allow the caller to specify that the cache should be ignored.
That would mean, the EID is recomputed while the new value is not stored in
cache.

BUG=chromium:835759
TEST=unit test and manual test with cryptohome --get_enrollment_id --recompute

Change-Id: Ifbad66fb2054fd12b7f4c6f89fa2657d6a8762d7
Reviewed-on: https://chromium-review.googlesource.com/1024031
Commit-Ready: Igor <igorcov@chromium.org>
Tested-by: Igor <igorcov@chromium.org>
Reviewed-by: Igor <igorcov@chromium.org>
Reviewed-by: Andrey Pronin <apronin@chromium.org>