cryptohome: Read VPD value faster and handle newlines better
Use vpd_get_value in order to obtain VPD values as this will use a
cache or the /sys/firmware/vpd filesystem if available. Note that this
also allows the value to be read from the RW VPD which is okay and even
desirable for testing.

To test, procure a device that has a TPM 1.2 (as TPM 2.0 doen't use
the same daemon). Ensure that there is a 64 hex character value for the
VPD key stable_device_secret_DO_NOT_SHARE (in either RO or RW VPD). If
not, create it as follows:

  # vpd -i RW_VPD -s stable_device_secret_DO_NOT_SHARE=$(openssl
      rand -hex 32) && reboot

Then verify that cryptohomed is running (the process number can be
any number):

  # status cryptohomed
  cryptohomed start/running, process 16098
  #

BUG=None
TEST=See description.

Change-Id: Ic2dafa76389e2c4dc16ff7a6aa239b30119ed9b6
Reviewed-on: https://chromium-review.googlesource.com/1024053
Commit-Ready: Yves Arrouye <drcrash@chromium.org>
Tested-by: Yves Arrouye <drcrash@chromium.org>
Reviewed-by: Andrey Pronin <apronin@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>