chromeos-config: Break up x86 identity impl
In order to implement ARM support and keep the ARM implementation
decoupled from the x86 implementation, the basic identity structure is
going to be split up.

This is the first in a series of changes to support this:
1) THIS PATCH
2) Change cros_config to pass identity object CrosConfigImpl as
CrosConfigIdentityX86.  This will allow the impl classes to have two
overrides that take the respective x86 or ARM identity impls.
3) Update tests to construct their own identity instances and pass to
CrosConfig for init.
4) Add new ARM based implementation of Identity and stitch in based on
the aforementioned refactoring changes.

BUG=chromium:832278
TEST=FEATURES=test emerge-reef chromeos-config-tools

Change-Id: I4be55183b6f5bbb1adac7166346068d84e027a0b
Reviewed-on: https://chromium-review.googlesource.com/1025010
Commit-Ready: C Shapiro <shapiroc@google.com>
Tested-by: C Shapiro <shapiroc@google.com>
Reviewed-by: Simon Glass <sjg@chromium.org>