chromeos-config: Add support for per-arch init
In order to implement ARM support and keep the ARM implementation
decoupled from the x86 implementation, the basic identity structure is
going to be split up.

This is the first in a series of changes to support this:
1) Break up the X86 and ARM identity implementations.
2) [THIS PATCH] Change cros_config to pass identity object CrosConfigImpl as
CrosConfigIdentityX86.  This will allow the impl classes to have two
overrides that take the respective x86 or ARM identity impls.
3) Update tests to construct their own identity instances and pass to
CrosConfig for init.
4) Add new ARM based implementation of Identity and stitch in based on
the aforementioned refactoring changes.

BUG=chromium:832278
TEST=FEATURES=test emerge-reef chromeos-config-tools

Change-Id: Ia7aa4cdcbfaa948bdf30e27038616e14bbda8e1e
Reviewed-on: https://chromium-review.googlesource.com/1024742
Commit-Ready: C Shapiro <shapiroc@google.com>
Tested-by: C Shapiro <shapiroc@google.com>
Reviewed-by: Simon Glass <sjg@chromium.org>