init: Fix _lsofFn_filter's sed script.
'x' was to pull the buffer *and* set the current line to the buffer.
Hence, for instance, if the output of lsof was

------
p123
n/foo
n/bar
------

it was transformed to

------
123
n/foo
------

because 'x' while processing n/foo replaces the buffer p123 with n/foo.

The new implementation using 'g' (that just pulls the buffer data) will yield

------
123
123
------

which works fine.

BUG=chromium:836350
TEST=`. /sbin/killers; kill_with_open_files_on /var` won't kill non-PIDs.

Change-Id: Ibc087aff28d96e370d314a58356282be712c6b80
Reviewed-on: https://chromium-review.googlesource.com/1027335
Commit-Ready: Kazuhiro Inaba <kinaba@chromium.org>
Tested-by: Kazuhiro Inaba <kinaba@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>