vm_tools: Enhance garcon icon searching when scale factor set
If garcon gets a request for an icon with a scale factor, but it can't
find one, then it will search for an icon with no scale factor and
return that instead. Currently I don't see any icons getting installed
with scale factors, and by default (at least on eve) we have a scale
factor of 2 and doing the icon scaling ourselves is better than no icons
at all.

BUG=chromium:836044
TEST=Verified garcon returns alternate icons

Change-Id: Ie674e97189cd94a3d69dd2126be7bec6911efabe
Reviewed-on: https://chromium-review.googlesource.com/1028585
Commit-Ready: Jeffrey Kardatzke <jkardatzke@google.com>
Tested-by: Jeffrey Kardatzke <jkardatzke@google.com>
Reviewed-by: Chirantan Ekbote <chirantan@chromium.org>