login, libchromeos-ui: Handle --enable-blink-features.
Update ChromiumCommandBuilder to expose a new
AddBlinkFeatureEnableOverride method for adding values to
--enable-blink-features and update session_manager to use
it. This flag contains a comma-separated list of values, so
I think we'd have the same problems as with --vmodule and
--enable-features if multiple AddArg calls tried to set it
-- Chrome will only honor one of the arguments.

Also update code in session_manager that was setting
--enable-features via AddArg (AddFeatureEnableOverride
should be used instead) and add DCHECKs to
ChromiumCommandBuilder to catch AddArg calls that
incorrectly add value-list flags.

BUG=none
TEST=updated unit tests pass; also added --enable-features
     and --enable-blink-features directives to
     /etc/chrome_dev.conf and verified that command line is
     constructed correctly

Change-Id: Ie7f79692816186addd96d59960442eeb09002c5f
Reviewed-on: https://chromium-review.googlesource.com/1029545
Commit-Ready: Dan Erat <derat@chromium.org>
Tested-by: Dan Erat <derat@chromium.org>
Reviewed-by: Mattias Nissler <mnissler@chromium.org>