vm_tools: Fixed icon path searching and size limit
These changes are required for VSCode's icon to show up. We were not
also searching /usr/share/pixmaps as a last resort which we should
according to the spec. We also had a very low limit set for max icon
size of 64k, that's now 1MB which is more reasonable.

BUG=chromium:836044
TEST=Verified VSCode icon shows up properly

Change-Id: I992459bbe2a79a8b41aab8f31840cd88458d6db0
Reviewed-on: https://chromium-review.googlesource.com/1031378
Commit-Ready: Jeffrey Kardatzke <jkardatzke@google.com>
Tested-by: Jeffrey Kardatzke <jkardatzke@google.com>
Reviewed-by: Chirantan Ekbote <chirantan@chromium.org>
Reviewed-by: Tim Zheng <timzheng@chromium.org>