shill: Move browser traffic uids from VpnProvider into Manager
This is in preparation of the Always On VPN feature which will
blackhole traffic for those uids (the users chronos and debugd).

For this, the Device class also needs access to the uids, so we
store the uids in the Manager, which can be accessed both from
VpnProvider and Device.

BUG=b:72861260
TEST=unit tests

Change-Id: Icdecdda5316eaf7ac9c9b698abc43b6ebd124971
Reviewed-on: https://chromium-review.googlesource.com/1032932
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Tested-by: Philipp Weiß <phweiss@chromium.org>
Reviewed-by: Ben Chan <benchan@chromium.org>