cryptohome: Fuzzer for CryptoLib::RsaOaepDecrypt
Add a fuzzer for the CryptoLib::RsaOaepDecrypt() function.

Also add a corpus with several valid inputs and a generator for the
corpus. This is essential because the main input of the tested
function is the RSA encrypted data, and it's unrealistic for the
fuzzer to brute force a valid pair of an RSA key and the encrypted
data.

BUG=chromium:837253
TEST=build and run fuzzer locally

Change-Id: I76667ccda3797bfa72669ea1dd559a818d87b03d
Reviewed-on: https://chromium-review.googlesource.com/1039750
Commit-Ready: Maksim Ivanov <emaxx@chromium.org>
Tested-by: Maksim Ivanov <emaxx@chromium.org>
Reviewed-by: Manoj Gupta <manojgupta@chromium.org>
Reviewed-by: Jonathan Metzman <metzman@chromium.org>