crosdns: Initial version of hostname resolver service
Design doc: go/crostini-hostnames

This adds a service where hostname:ip mappings can be set and removed.
Currently this will write to the /etc/hosts file in order to accomplish
this. Later this service can be expanded to cover additional name
resolution/DNS functionality.

This is the complete code for the first version, with the exception of
the seccomp policy which will be done after the code is approved for the
service.

BUG=chromium:825010
TEST=Unit tests pass, verified with dbus-send on Chromebook
CQ-DEPEND=CL:1045140

Change-Id: I97c8b471d4f156a33e0392eeee55d96d49488781
Reviewed-on: https://chromium-review.googlesource.com/1045250
Commit-Ready: Jeffrey Kardatzke <jkardatzke@google.com>
Tested-by: Jeffrey Kardatzke <jkardatzke@google.com>
Reviewed-by: Ben Chan <benchan@chromium.org>