container_utils: disambiguate file descriptor ownership
The DeviceJailServer constructor takes ownership of the passed
file descriptor, but the FD is passed bare, meaning it might
be interpreted as a "weak" reference to the FD that it doesn't
clean up. Make the factory method instantiate a ScopedFD instead
and move it into the DeviceJailServer when that's constructed.

BUG=None
TEST=emerge, deploy, use device_jail_utility --server

Change-Id: I83c2ac5c93ea9255825f573d080369c50bf6f92e
Reviewed-on: https://chromium-review.googlesource.com/1045552
Commit-Ready: Eric Caruso <ejcaruso@chromium.org>
Tested-by: Eric Caruso <ejcaruso@chromium.org>
Reviewed-by: Ben Chan <benchan@chromium.org>