cryptohome: Improve LE Credential free label finding scheme
This patch modifies the GetFreeLabel() to generate a random new
label instead of just a monotonically increasing value.

This ensures that the labels are spread more evenly across the entire
hash tree. This helps in a out-of-sync situation where the first label
being out of sync would always make the entire tree out-of-sync.

This patch also adds a unit test to test the change required in
PersistentLookupTable.

BUG=chromium:809718, chromium:809703
TEST=1. cros_run_unit tests
     2. Make several new keys using CLI, and verify that the key
     labels in the low_entropy_creds directory are random.

Change-Id: I017a75b657b710c74a1868aa14eca58fcd9e3363
Reviewed-on: https://chromium-review.googlesource.com/1045503
Commit-Ready: Prashant Malani <pmalani@google.com>
Tested-by: Prashant Malani <pmalani@google.com>
Reviewed-by: Prashant Malani <pmalani@google.com>