chromeos-config: Implement merge support for YAML
Added support so that YAML snippets can be used to override configs.
This feature can either match (fully/partially) the identity attributes
present in the final config and override based on that or it can be used
to simply append new configs (if folks want to keep the YAML source
split up).

BUG=b:72807610
TEST=unit

Change-Id: Ie578c456c28f5dfd9cde6fa5622b512a183921f9
Reviewed-on: https://chromium-review.googlesource.com/1048388
Commit-Ready: C Shapiro <shapiroc@google.com>
Tested-by: C Shapiro <shapiroc@google.com>
Reviewed-by: Lann Martin <lannm@chromium.org>
Reviewed-by: Gregory Meinke <gmeinke@chromium.org>
Reviewed-by: C Shapiro <shapiroc@google.com>