shill: Add programatic support for minijail '-G' flag to ProcessManager
This CL is in preparation to run the shill process tree as a non-root
user.

Shill will soon spawn openvpn as user/group 'shill' before openvpn drops to
user/group 'openvpn'. Startup process for openvpn needs write access
to /dev/net/tun, which requires making shill user part of the 'tun'
group and allowing the jailed openvpn process to inherit the
supplementary groups from the 'shill' UID before dropping to 'openvpn'.

Unfortunately, adding this single flag pushed the number of args in the
ProcessManager:SpawnProcessInMinijailWithPipes function to 11, which is
effectively prevented by gMock code, which only allows for mocking
functions that use up to 10 arguments. Hence, this code also refactors
the function to take in a struct with the 3 standard file descriptors
instead of passing them each individually as arguments.

BUG=chromium:649417
TEST=unittests pass, tested as part of larger sandbox shill debug CL

Change-Id: I8254329cbe34de6551688b93d319f15489c3817b
Reviewed-on: https://chromium-review.googlesource.com/1050730
Commit-Ready: Micah Morton <mortonm@chromium.org>
Tested-by: Micah Morton <mortonm@chromium.org>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>