vm_tools: Add SSH keys to vm_concierge
This adds SSH key generation, retrieval and passing to the container for
the purpose of SFTP mounting the container's filesystem in the Chrome OS
Files app.  There is one keypair for the host and then one keypair per
container.  Container keys are deleted if their corresponding VM is
deleted.

BUG=chromium:832508
TEST=Verified keys are generated and passed into run_container script

Change-Id: I8818bf6c64294d83750bdcfb4ef84e028ec0dd2d
Reviewed-on: https://chromium-review.googlesource.com/1053296
Commit-Ready: Jeffrey Kardatzke <jkardatzke@google.com>
Tested-by: Jeffrey Kardatzke <jkardatzke@google.com>
Reviewed-by: Stephen Barber <smbarber@chromium.org>