trunks: Fix uninitialized read in unit test
This fixes the failure of the
TpmUtilityTest.GetPublicRSAEndorsementKeyModulus_NoDataInNvram
test under ASan.

The failure was caused by the read of uninitialized integer
TPMS_NV_PUBLIC.data_size in
TpmUtilityImpl::GetPublicRSAEndorsementKeyModulus(). Normally
this field is initialized by Tpm::NV_ReadPublicSync(), but the
test's default mock behavior was returning TPM_RC_SUCCESS
without touching the returned structure. The rest of the test,
consequently, falls under undefined behavior - but it was
failing only if the uninitialized integer turns out to be large
enough to bail out from ReadNVSpace() with SAPI_RC_BAD_SIZE
(which was happening under ASan).

The fix is to provide the missing mock, so that the Tpm method
doesn't report success without setting the data.

BUG=chromium:841571
TEST=Run the GetPublicRSAEndorsementKeyModulus_NoDataInNvram
     test under USE=asan

Change-Id: I7f94fdc768969c4831b8fee9acf537e952d6e754
Reviewed-on: https://chromium-review.googlesource.com/1053447
Commit-Ready: Maksim Ivanov <emaxx@chromium.org>
Tested-by: Maksim Ivanov <emaxx@chromium.org>
Reviewed-by: Igor <igorcov@chromium.org>
Reviewed-by: Andrey Pronin <apronin@chromium.org>