power: Don't report missing current or voltage as 0.
If the current_now or voltage_now sysfs nodes are missing
from a line power supply, make powerd omit them from "On AC"
log messages and make power_supply_info print "unknown". We
formerly treated missing nodes as 0, making it impossible to
tell whether that was the actual value that was seen or not.

BUG=chromium:807753
TEST=updated unit tests; also verified that caroline now
     logs the following instead of a 0 current:
     "On AC (USB_PD, 14.6V, max 2.0A at 15.0V) with ..."
     and that power_supply_info prints:
     "current (A):             unknown"

Change-Id: I8bd182ded113e3d677dfe092fdf53ed32401f5a9
Reviewed-on: https://chromium-review.googlesource.com/1055790
Commit-Ready: Dan Erat <derat@chromium.org>
Tested-by: Dan Erat <derat@chromium.org>
Reviewed-by: Brian Norris <briannorris@chromium.org>