biod: return accurate record in CrosFpBiometricsManager
Use the newly added match-index field in the cros_fp MKBP event in order
to be always able to return the accurate finger record on a successful
match.
Remove the previous workarounds.

This requires an up-to-date FPMCU firmware with crrev.com/c/1047267.
With previous firmware versions, this will always return 'finger 0'.

BUG=b:77516790
TEST=On Meowth, enroll several fingers, then check that the fingerprint
matching still work at the unlock screen (and the proper template index
is printed /var/log/biod/biod.LATEsT). Go to the Fingerprint settings
page and see the proper finger highlighted in the list when touching the
sensor.

Change-Id: Iece1c601cffe3029df8706969d2e5e2020ae192d
Reviewed-on: https://chromium-review.googlesource.com/1054871
Commit-Ready: Vincent Palatin <vpalatin@chromium.org>
Tested-by: Vincent Palatin <vpalatin@chromium.org>
Reviewed-by: Nicolas Norvez <norvez@chromium.org>