biod: fix record removal in CrosFpBiometricsManager
Ok but not great solution to fix the biod crash / half-baked implementation
when a finger template is removed (and it is not the last one).
As the FP MCU array of templates needs to stay synchronized with the
CrosFpBiometricsManager vision of it (stored as record_ vector), just
drop the current one and reload everything.

BUG=chromium:836435
TEST=On Meowth, enroll 2 fingers, then remove the first one from the
Fingerprint Settings UI: no longer crashing, first one no longer unlocks
the machine, the second one still unlocks the machine.

Change-Id: I4b87f641ec66d5f3932ef8ac02e29c3fbc50e889
Reviewed-on: https://chromium-review.googlesource.com/1054236
Commit-Ready: Vincent Palatin <vpalatin@chromium.org>
Tested-by: Vincent Palatin <vpalatin@chromium.org>
Reviewed-by: Nicolas Norvez <norvez@chromium.org>