chromeos-config: Open up power to all options
There are 40+ power settings and the yaml schema is really not the
source of truth for these settings.  For now at least (maybe we'll
revisit when there is more usage in yaml), we're going to allow any
setting in the yaml and point users to the c++ header explaining the
details of the respective properties.

BUG=None
TEST=unit

Change-Id: I17d7c88efb2e81911fa318caf4a5f6c06d28dc26
Reviewed-on: https://chromium-review.googlesource.com/1055472
Commit-Ready: C Shapiro <shapiroc@google.com>
Tested-by: C Shapiro <shapiroc@google.com>
Reviewed-by: C Shapiro <shapiroc@google.com>