cryptohome: Move existence check into HomeDirs
Move the code that checks cryptohome existence from the Mount class
into the HomeDirs class.

This refactoring allows to choose between the new cryptohome
creation and the existing cryptohome mounting in advance, without
the need to obtain an instance of Mount (which is conceptually
associated with already mounted cryptohome). This will be
necessary for implementing the challenge-response cryptohome
mounts, as for them the flow needs to be known before the
asynchronous challenge-response dance happens.

BUG=chromium:842791
TEST=existing unit tests (homedirs_unittest.cc, mount_unittest.cc)

Change-Id: Ic0802afa6074c425706832c4ed1044645af77624
Reviewed-on: https://chromium-review.googlesource.com/1057630
Commit-Ready: Maksim Ivanov <emaxx@chromium.org>
Tested-by: Maksim Ivanov <emaxx@chromium.org>
Reviewed-by: Ryo Hashimoto <hashimoto@chromium.org>