cryptohome: Extract obfuscated username building
Extract obfuscated username building code from the
UsernamePasskey class into a standalone function.

This allows to write a code that operates with the obfuscated
username before the passkey, needed for building an instance of
Credentials, is obtained.

This CL will be followed up with CLs that will replace the uses
of Credentials with the use of the obfuscated_username string,
where that's possible.

BUG=chromium:842791
TEST=new unit test (ObfuscatedUsername.*),
     existing unit tests (UsernamePasskeyTest.*)

Change-Id: Iba3167ad581d5e7f6325e28a3df0786f05df4ed6
Reviewed-on: https://chromium-review.googlesource.com/1057272
Commit-Ready: Maksim Ivanov <emaxx@chromium.org>
Tested-by: Maksim Ivanov <emaxx@chromium.org>
Reviewed-by: Ryo Hashimoto <hashimoto@chromium.org>
Reviewed-by: Andrey Pronin <apronin@chromium.org>