chromeos-config: Use the kernel driver instead of smbios tables
Now that the kernel provides a driver for accessing the SMBIOS name and
SKU ID, we can use that instead of decoding the binary SMBIOS tables.

This removes a significant amount of code and complexity. It also makes
x86 work in a similar way to ARM, in that it can just look at files
provided by the kernel.

Also drop creation of the SMBIOS file used by this code.

BUG=chromium:795122
TEST=FEATURES=test sudo -E emerge --nodeps chromeos-config-tools
Also manual test on santa with /var/run/cros_config/SMBIOS removed

Change-Id: Idb78f5568b3c2cbe483859895cb642dfbaa8f04b
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1062212
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Reviewed-by: Jason Clinton <jclinton@chromium.org>