init: Conditionally enables Quota for ext4
There are 2 parts:
1. For fresh installation, chromeos-install will mkfs the ext4
filesystem with quota option on.
2. Otherwise, chromeos_startup conditionally checks whether the quota
option is turned on. If it hasn't, the scripts turn it on.

BUG=b:62995196
TEST=- Turn off quota and checked if chromeos_startup enables it.
TEST=- Checked that chromeos_startup doesn't trigger tune2fs when quota
is on (by adding an else in the chromeos_startup change - and make sure
that the else is triggered).
TEST=- Turn off kernel quota config, and /mnt/stateful_partition is
still correctly mounted, without quota.

Change-Id: I7e62c7dd79ec65ec380b8049e2d77fd0778844da
Reviewed-on: https://chromium-review.googlesource.com/1064571
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Tested-by: Risan <risan@chromium.org>
Reviewed-by: Ryo Hashimoto <hashimoto@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>