metrics: memd: remove regex dependency
env_logger by default depends on regex, which depends on
a lot of other packages.  This reduces the size of the executable
from 1.4 to 0.6 MB.

This also accidentally bumped the version number for a number
of other dependencies.  Maybe I should freeze them all in Cargo.toml.

BUG=chromium:729335
TEST=tested upstart with start/stop/reboot

Change-Id: I3d94c35c50e4cae1b2853b06665fc6415b2a9166
Reviewed-on: https://chromium-review.googlesource.com/1063041
Commit-Ready: Luigi Semenzato <semenzato@chromium.org>
Tested-by: Luigi Semenzato <semenzato@chromium.org>
Reviewed-by: Luigi Semenzato <semenzato@chromium.org>
Reviewed-by: Zach Reizner <zachr@chromium.org>