cecservice: improve CEC compliance
I have run cec-compliance tool from https://linuxtv.org/wiki/index.php/V4l-utils
against cecservice. The tool complained about a couple of issues, this patch
adds small tweaks to make cec-compliance tool more happy:
1. Set OSD name on the device (that does not seem to be required by
the standard, but the tool complains if we don't have it set and it is
an easy fix).
2. Instead of responding to standby request with 'feature abort' we
simply ignore it (as per spec). Note - with that change the tool still
complains that we haven't gone to standby while instructed to do so,
but IMO that is OK as the specification allows for devices like
for example PCs to ignore standby requests.

With the patch applied the only error reported by the tool is about
cecservice ignoring standby requests.

BUG=chromium:828820
TEST=Run unit test and run cec-compliance tool.

Change-Id: I5cb6fae6e863e0f34f76440811ad44867c539086
Reviewed-on: https://chromium-review.googlesource.com/1065739
Commit-Ready: Dariusz Marcinkiewicz <darekm@google.com>
Tested-by: Dariusz Marcinkiewicz <darekm@google.com>
Reviewed-by: Felix Ekblom <felixe@chromium.org>
Reviewed-by: Ben Chan <benchan@chromium.org>