permission_broker: Add Firewall fuzzer.
The firewall code in PermissionBroker doesn't do a lot of parsing, but
it is exposed to possibly untrusted input from Chrome. Besides, it's a
good example to show how to fuzz an API that does more than just
consume a |char*|.

Implement a simple fake that short-circuits the calls to real
executables. An obvious improvement would be to fail some of those
calls. Use FuzzedDataProvider to exercise the API, and open and close
ports.

BUG=chromium:845891
TEST=Run the fuzzer.

Change-Id: I5c0ff8559c1c00e970e9949b11a0ed8cd5252c40
Reviewed-on: https://chromium-review.googlesource.com/1066574
Commit-Ready: Jorge Lucangeli Obes <jorgelo@chromium.org>
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Tested-by: Jorge Lucangeli Obes <jorgelo@chromium.org>
Reviewed-by: Manoj Gupta <manojgupta@chromium.org>
Reviewed-by: Jonathan Metzman <metzman@chromium.org>