vm_tools: Add rate limiting for OpenUrl call
This uses a fixed window rate limiting approach where the window timer
is reset when a request is received that is out of the current window. A
counter is increased each time a request is received while still in that
window and if we exceed our limit we start rejecting requests.

BUG=chromium:841610
TEST=Verified with manual testing

Change-Id: Ic7530c34b0140284d01864718bbfffc342ce99a3
Reviewed-on: https://chromium-review.googlesource.com/1065116
Commit-Ready: Jeffrey Kardatzke <jkardatzke@google.com>
Tested-by: Jeffrey Kardatzke <jkardatzke@google.com>
Reviewed-by: Chirantan Ekbote <chirantan@chromium.org>