rockchip: always qbuf for every enabled video node
To make sure the stream buffers of a request are written with processed
contents from the same MIPI frame, we always queue a buffer for every
enabled active node when processing every capture request.  Together
with the lockstep in the ISP kernel driver it would guarantee the
contents of all the buffers in the same request are consistent.

BUG=b:78779539
TEST=Make sure camera works on Dru.

Change-Id: Ie6ffda5bfc6ccda5d745ac1a9a60b6b1bb8716e3
Reviewed-on: https://chromium-review.googlesource.com/1066251
Commit-Ready: Ricky Liang <jcliang@chromium.org>
Tested-by: Ricky Liang <jcliang@chromium.org>
Reviewed-by: Hung-yu Wu <hywu@chromium.org>