cryptohome: Bind TPM 1.2 delegate to PCR0
Make the TPM 1.2 delegate be bound to PCR0. This will act as
an additional measure of protection against switches between
verified and dev modes, as this results in the change of PCR
number 0.

The exact effect is that if the current PCR0 value is
different from the one when the delegate was created, such
operations like the attestation or the migration of a CMK
(Certified Migratable Key) will stop working.

BUG=chromium:806788
TEST=existing tests

Change-Id: Id347b911d102e6bf2795441e326e8ee2a427e114
Reviewed-on: https://chromium-review.googlesource.com/1065826
Commit-Ready: Maksim Ivanov <emaxx@chromium.org>
Tested-by: Maksim Ivanov <emaxx@chromium.org>
Reviewed-by: Andrey Pronin <apronin@chromium.org>