arc-setup: Mount demo apps for demo sessions
When setting up ARC container for demo sessions (during continue boot),
mount demo apps to the shared mount passed to the ARC container.
Demo apps will be part of demo mode resources image, which is a
dm-verity image kept on the devices stateful partition, and intended to
be loaded by imageloader service.
The file path from which demo apps should be loaded (which is also
expected to be a squashfs image) will be passed as demo_session_apps
parameter of the arc container upgrade request (which translates to
DEMO_SESSION_APPS env variable) - arc setup will confirm that this path
is a descendant of /run/imageloader directory.

Also, pass --mount-demo-apps to arcbootcontinue, so it knows to mount
the demo apps (the demo apps will be mounted as /odm patition).

CQ-DEPEND=CL:1066498
BUG=chromium:842979, b:79268228

TEST=With Chrome patched to set is_demo_session for UpgradeArcContainer
request, verify that ARC container mounts demo apps on odm

Change-Id: Iaac4e1bd90635042dbe2c66c0e042a2c28299722
Reviewed-on: https://chromium-review.googlesource.com/1066284
Commit-Ready: Toni Barzic <tbarzic@chromium.org>
Tested-by: Toni Barzic <tbarzic@chromium.org>
Reviewed-by: Yusuke Sato <yusukes@chromium.org>