metrics: memd: fix unwraps
I noticed I still had a couple of incorrect unwrap() calls.
I changed one to ?, the others to expect().  There are still
another couple of unwraps left, but those should be OK because
they are expected to succeed---it's an internal bug if they fail.

BUG=chromium:729335
TEST=none

Change-Id: I93ceb611fb82ff862bec08b59d69dcffe863a222
Reviewed-on: https://chromium-review.googlesource.com/1068129
Commit-Ready: Luigi Semenzato <semenzato@chromium.org>
Tested-by: Luigi Semenzato <semenzato@chromium.org>
Reviewed-by: Stephen Barber <smbarber@chromium.org>
Reviewed-by: Zach Reizner <zachr@chromium.org>