chromeos-config: support to query property by assigning the SKU ID.
Currrently cros_config only allows you to query the property in
the SKU ID of this device if test database is not passed in. But
in some use cases, we would need to do query for a SKU ID other then
this device belonged to and use the default database in this device.

BUG=chromium:838134
TEST=build new image and manually test in the DUT

Change-Id: Ia57562b3d4082285838b8f67b847d67cc615918e
Reviewed-on: https://chromium-review.googlesource.com/1065721
Commit-Ready: Marco Chen <marcochen@chromium.org>
Tested-by: Marco Chen <marcochen@chromium.org>
Reviewed-by: Marco Chen <marcochen@chromium.org>