cryptohome: Decide on ephemeral mount in Service
Extract the logic of deciding on ephemeral mount from the
Mount class into the Service class.

This refactoring allows to know earlier, before calling
into Mount, whether the mount will be ephemeral - so that
we will be able to know in case of challenge-response
credentials that we need neither to unseal credentials
from an existing serialized vault keyset nor to generate
fresh credentials.

BUG=chromium:842791
TEST=existing unit tests

Change-Id: Id884bb75d4e9083537967ca71e071a135a9ddaa4
Reviewed-on: https://chromium-review.googlesource.com/1065684
Commit-Ready: Maksim Ivanov <emaxx@chromium.org>
Tested-by: Maksim Ivanov <emaxx@chromium.org>
Reviewed-by: Ryo Hashimoto <hashimoto@chromium.org>