Pipe SUPERVISION_TRANSITION info to ARC on boot.
This setting informs ARC that a user has logged in with a different
account type, and needs to have supervision enabled or disabled. This
happens when a child account graduates to regular account, or when some
accounts opt-in to being supervised.

Design doc: go/chromeos-usm-arc-dd

BUG=chromium:836246
TEST=FEATURES=test emerge-${BOARD} chromeos-login

Change-Id: Ife5a2f873d9f7481bbeb3b622b247750f2323273
Reviewed-on: https://chromium-review.googlesource.com/1069031
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Tested-by: Luis Hector Chavez <lhchavez@chromium.org>
Reviewed-by: Luis Hector Chavez <lhchavez@chromium.org>
Reviewed-by: Yury Khmel <khmel@chromium.org>