cryptohome: Split LE hash cache
To facilitate a clean Log replay functionality in the case of the
on-disk Low Entropy (LE) credentials state being out of sync with
Cr50, this CL splits the HashCache into two parts:
- LeafCache: A memory mapped persistent file which stores the MAC
  values for all leaf labels.
- InnerHashArray: An in-memory array storing the hash values of
  all inner nodes of the Hash Tree.

This way, the InnerHashArray will always be generated from the
LeafCache on the first operation (or during LE initialization after
boot). If any discrepancies are there, we will know for sure that
the LeafCache (and concordantly, our on-disk tree state) is erroneous
and we can begin a recovery operation to fix it.

BUG=chromium:794010, chromium:809710
TEST=cryptohome unit tests pass.

Change-Id: I3678d41268c4789e88fd2b44e838b2d001c154af
Reviewed-on: https://chromium-review.googlesource.com/1069838
Commit-Ready: Prashant Malani <pmalani@google.com>
Tested-by: Prashant Malani <pmalani@google.com>
Reviewed-by: Prashant Malani <pmalani@google.com>