vm_tools: Add D-Bus and gRPC service to cicerone
This is a refactoring of concierge into cicerone. The first patch set
uploaded is exact copies of container_listener_impl.*, service.* and
virtual_machine* from concierge. This should make this much easier to
review since the vast majority of these changes are either removals or
copies from concierge. The additional D-Bus service functionality needed
in cicerone that concierge uses for communication is also included here.

This CL also disables starting of the container_listener service inside
of cicerone, yet leaves all that functionality there. This is to keep
the delta from concierge to a minimum. The next CL will remove all of
the corresponding pieces from concierge (one of which is a full removal
of container_listener) and in that CL I will turn on container_listener
in cicerone so we don't have a port conflict.

BUG=chromium:842845
TEST=Verified with full set of code that this works on eve
CQ-DEPEND=CL:1069767

Change-Id: I5fc335b58181cc055b9279b3eb52085428a19894
Reviewed-on: https://chromium-review.googlesource.com/1070871
Commit-Ready: Jeffrey Kardatzke <jkardatzke@google.com>
Tested-by: Jeffrey Kardatzke <jkardatzke@google.com>
Reviewed-by: Stephen Barber <smbarber@chromium.org>