cryptohome: Fix lockbox validity check in mount-encrypted
mount-encrypted only considered the lockbox space valid when the TPM
is owned. For stateful preservation, this will not necessarily be the
case as we only take ownership if the encstateful NVRAM space must be
defined.

To decide lockbox validity properly, add a CheckLockbox function in
SystemKeyLoader, so the TPM 1.2 implementation can take into account
presence of a lockbox MAC in the encstateful space and declare the
lockbox valid in case the MAC validates, even for the case where the
TPM isn't owned yet.

BUG=chromium:845885
TEST=Updated unit tests.

Change-Id: Ib75a9d765e3dc50685851c96cdacb0380583fb69
Reviewed-on: https://chromium-review.googlesource.com/1071528
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Tested-by: Mattias Nissler <mnissler@chromium.org>
Reviewed-by: Andrey Pronin <apronin@chromium.org>