authpolicy: Ping server if TGT is not valid
Currently, if the TGT is not valid as the user enters a session (e.g.
after an offline login), Chrome shows a popup that the user has to log
out and back in to get a new TGT. However, if the Active Directory
server is unavailable, this fails and Chrome shows the popup again.

This CL fixes this issue by pinging the server in GetUserStatus() if the
TGT is not valid. If the ping fails, GetUserStatus() returns
ERROR_NETWORK_PROBLEM and Chrome ignores the result and tries again
an hour later.

Also reworded some comments that seemed to indicate that
AuthenticateUser() was never called, even though it was (and failed).

BUG=chromium:844662
TEST=On an AD enrolled device, log in while offline.
     Chrome should NOT show the popup on login.

Change-Id: I07920dcfcb6bcf5a05ec6d4de29bf57e4efedc1a
Reviewed-on: https://chromium-review.googlesource.com/1072470
Commit-Ready: Lutz Justen <ljusten@chromium.org>
Tested-by: Lutz Justen <ljusten@chromium.org>
Reviewed-by: Roman Sorokin <rsorokin@chromium.org>