cros-disks: Get the realpath of the parent if the datadir doesn't exist.
When mounting DriveFS, on the first mount for a particular datadir,
cros-disks is responsible for creating that directory. Currently, it
attempts to find the realpath of the datadir, which fails before it has
been created. When the datadir doesn't exist, find the realpath of the
parent directory instead.

BUG=chromium:829274
TEST=unit tests

Change-Id: I3daf27f69ebd85d18f5a47cbe31033d8255e39e3
Reviewed-on: https://chromium-review.googlesource.com/1071151
Commit-Ready: Sam McNally <sammc@chromium.org>
Tested-by: Sam McNally <sammc@chromium.org>
Reviewed-by: Ben Chan <benchan@chromium.org>